+++ apps/backoffice-v2/src/lib/blocks/components/KycBlock/hooks/useKycBlock/useKycBlock.tsx
@@ -21,6 +21,10 @@ import { MotionBadge } from '../../../../../../common/components/molecules/Motio
 import { capitalize } from '../../../../../../common/utils/capitalize/capitalize';
 import { useStorageFilesQuery } from '../../../../../../domains/storage/hooks/queries/useStorageFilesQuery/useStorageFilesQuery';
 import { TWorkflowById } from '../../../../../../domains/workflows/fetchers';
+import { useToggle } from '@/common/hooks/useToggle/useToggle';
+import { generateEditableDetailsV2Fields } from '@/common/components/organisms/EditableDetailsV2/utils/generate-editable-details-v2-fields';
+import { useUpdateContextAndSyncEntityMutation } from '@/domains/workflows/hooks/mutations/useUpdateContextAndSyncEntity/useUpdateContextAndSyncEntity';
+import { useEventMutation } from '@/domains/workflows/hooks/mutations/useEventMutation/useEventMutation';
 
 const motionBadgeProps = {
   exit: { opacity: 0, transition: { duration: 0.2 } },
@@ -175,15 +179,6 @@ export const useKycBlock = ({
       ) ?? []
     : [];
 
-  const details = Object.entries(childWorkflow?.context?.entity?.data ?? {}).map(
-    ([title, value]) => ({
-      title,
-      value,
-      pattern: '',
-      isEditable: false,
-      dropdownOptions: undefined,
-    }),
-  );
   const documents = childWorkflow?.context?.documents?.flatMap(
     (document, docIndex) =>
       document?.pages?.map(({ type, metadata, data }, pageIndex) => ({
@@ -308,9 +303,8 @@ export const useKycBlock = ({
               }}
               disabled={isDisabled}
               size={'wide'}
-              className={ctw({
-                '!bg-success': !isDisabled,
-              })}
+              variant={'success'}
+              className={'enabled:bg-success enabled:hover:bg-success/90'}
             >
               Approve
             </MotionButton>
@@ -341,13 +335,32 @@ export const useKycBlock = ({
       .flat(1);
   };
 
+  const { mutate: mutateInitiateKyc } = useEventMutation();
+
+  const getEvent = () => {
+    if (childWorkflow?.nextEvents?.includes('start')) {
+      return 'start';
+    }
+  };
+  const event = getEvent();
+  const onInitiateKyc = useCallback(() => {
+    if (!event) {
+      return;
+    }
+
+    mutateInitiateKyc({
+      workflowId: childWorkflow?.id,
+      event,
+    });
+  }, [mutateInitiateKyc, event, childWorkflow?.id]);
+
   const headerCell = createBlocksTyped()
     .addBlock()
     .addCell({
       id: 'header',
       type: 'container',
       props: {
-        className: 'items-start',
+        className: 'justify-between items-center pt-6',
       },
       value: createBlocksTyped()
         .addBlock()
@@ -356,11 +369,125 @@ export const useKycBlock = ({
           value: `${valueOrNA(childWorkflow?.context?.entity?.data?.firstName)} ${valueOrNA(
             childWorkflow?.context?.entity?.data?.lastName,
           )}`,
+          props: {
+            className: 'mt-0',
+          },
+        })
+        .addCell({
+          type: 'callToAction',
+          value: {
+            text: 'Initiate KYC',
+            onClick: onInitiateKyc,
+            props: {
+              className:
+                'justify-self-end px-2 py-0 text-xs aria-disabled:pointer-events-none aria-disabled:opacity-50',
+              variant: 'outline',
+              disabled: !event,
+            },
+          },
         })
         .build()
         .flat(1),
     })
     .cellAt(0, 0);
+
+  const fields = generateEditableDetailsV2Fields(childWorkflow?.context)({
+    path: 'entity.data',
+  });
+
+  const [isEditable, _toggleIsEditable, toggleOnIsEditable, toggleOffIsEditable] = useToggle();
+  const { mutate: mutateUpdateContextAndSyncEntity } = useUpdateContextAndSyncEntityMutation({
+    workflowId: childWorkflow?.id,
+    onSuccess: () => {
+      toggleOffIsEditable();
+    },
+  });
+
+  const onSubmit = useCallback(
+    (values: Record<PropertyKey, any>) => {
+      mutateUpdateContextAndSyncEntity(values);
+    },
+    [mutateUpdateContextAndSyncEntity],
+  );
+
+  const getEntityDataBlock = () => {
+    if (parentWorkflow?.workflowDefinition?.config?.editableContext?.kyc?.entity) {
+      return createBlocksTyped()
+        .addBlock()
+        .addCell({
+          type: 'editableDetails',
+          value: fields,
+          props: {
+            title: 'Details',
+            onSubmit,
+            onEnableIsEditable: toggleOnIsEditable,
+            onCancel: toggleOffIsEditable,
+            config: {
+              parse: {
+                date: true,
+                isoDate: true,
+                datetime: true,
+                boolean: true,
+                url: true,
+                nullish: true,
+              },
+              blacklist: [],
+              actions: {
+                options: {
+                  disabled: !caseState.writeEnabled,
+                },
+                enableEditing: {
+                  disabled: isEditable,
+                },
+                editing: {
+                  disabled: !isEditable || !caseState.writeEnabled,
+                },
+                cancel: {
+                  disabled: false,
+                },
+                save: {
+                  disabled: !caseState.writeEnabled,
+                },
+              },
+              inputTypes: {
+                dateOfBirth: 'date',
+              },
+            },
+          },
+        })
+        .build()
+        .flat(1);
+    }
+
+    return createBlocksTyped()
+      .addBlock()
+      .addCell({
+        id: 'header',
+        type: 'heading',
+        value: 'Details',
+      })
+      .addCell({
+        id: 'decision',
+        type: 'details',
+        value: {
+          id: 1,
+          title: 'Details',
+          data: Object.entries(childWorkflow?.context?.entity?.data ?? {}).map(
+            ([title, value]) => ({
+              title,
+              value,
+              pattern: '',
+              isEditable: false,
+              dropdownOptions: undefined,
+            }),
+          ),
+        },
+        workflowId: childWorkflow?.id,
+        documents: childWorkflow?.context?.documents,
+      })
+      .build()
+      .flat(1);
+  };
 
   return createBlocksTyped()
     .addBlock()
@@ -409,63 +536,74 @@ export const useKycBlock = ({
                 .addBlock()
                 .addCell({
                   type: 'container',
-                  value: createBlocksTyped()
-                    .addBlock()
-                    .addCell({
-                      id: 'header',
-                      type: 'heading',
-                      value: 'Details',
-                    })
-                    .addCell({
-                      id: 'decision',
-                      type: 'details',
-                      value: {
-                        id: 1,
-                        title: 'Details',
-                        data: details,
-                      },
-                      workflowId: childWorkflow?.id,
-                      documents: childWorkflow?.context?.documents,
-                    })
-                    .build()
-                    .flat(1),
+                  value: getEntityDataBlock(),
                 })
                 .addCell({
                   type: 'container',
-                  value: createBlocksTyped()
-                    .addBlock()
-                    .addCell({
-                      id: 'header',
-                      type: 'heading',
-                      value: 'Document Extracted Data',
-                    })
-                    .build()
-                    .concat(documentExtractedData)
-                    .flat(1),
+                  value: documentExtractedData.length
+                    ? createBlocksTyped()
+                        .addBlock()
+                        .addCell({
+                          id: 'header',
+                          type: 'heading',
+                          value: 'Document Extracted Data',
+                        })
+                        .build()
+                        .concat(documentExtractedData)
+                        .flat(1)
+                    : createBlocksTyped()
+                        .addBlock()
+                        .addCell({
+                          type: 'heading',
+                          value: 'Document Extracted Data',
+                        })
+                        .addCell({
+                          type: 'paragraph',
+                          value: 'Initiate KYC for document extracted data to appear',
+                          props: {
+                            className: 'py-4 text-slate-500',
+                          },
+                        })
+                        .buildFlat(),
                 })
                 .addCell({
                   type: 'container',
-                  value: createBlocksTyped()
-                    .addBlock()
-                    .addCell({
-                      id: 'header',
-                      type: 'heading',
-                      value: 'Document Verification Results',
-                    })
-                    .addCell({
-                      id: 'decision',
-                      type: 'details',
-                      hideSeparator: true,
-                      value: {
-                        id: 1,
-                        title: 'Decision',
-                        data: decision,
-                      },
-                      workflowId: childWorkflow?.id,
-                      documents: childWorkflow?.context?.documents,
-                    })
-                    .build()
-                    .flat(1),
+                  value: decision.length
+                    ? createBlocksTyped()
+                        .addBlock()
+                        .addCell({
+                          id: 'header',
+                          type: 'heading',
+                          value: 'Document Verification Results',
+                        })
+                        .addCell({
+                          id: 'decision',
+                          type: 'details',
+                          hideSeparator: true,
+                          value: {
+                            id: 1,
+                            title: 'Decision',
+                            data: decision,
+                          },
+                          workflowId: childWorkflow?.id,
+                          documents: childWorkflow?.context?.documents,
+                        })
+                        .build()
+                        .flat(1)
+                    : createBlocksTyped()
+                        .addBlock()
+                        .addCell({
+                          type: 'heading',
+                          value: 'Document Verification Results',
+                        })
+                        .addCell({
+                          type: 'paragraph',
+                          value: 'Initiate KYC for document verification results to appear',
+                          props: {
+                            className: 'py-4 text-slate-500',
+                          },
+                        })
+                        .buildFlat(),
                 })
                 .build()
                 .flat(1),