+++ apps/backoffice-v2/src/lib/blocks/hooks/useDirectorsBlocks/useDirectorsBlocks.tsx
@@ -253,6 +253,7 @@ export const useDirectorsBlocks = ({
                           disabled={isApproveDisabled}
                           size={'wide'}
                           variant={'success'}
+                          className={'enabled:bg-success enabled:hover:bg-success/90'}
                         >
                           Approve
                         </MotionButton>