+++ apps/backoffice-v2/src/tests-setup.ts
@@ -1,4 +1,1 @@
-import { expect } from 'vitest';
-import matchers from '@testing-library/jest-dom/matchers';
-
-expect.extend(matchers);
+import '@testing-library/jest-dom/vitest';