+++ services/workflows-service/jest.config.cjs
@@ -1,6 +1,7 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  testTimeout: 30000,
   modulePathIgnorePatterns: ['<rootDir>/dist/'],
   testRegex: '(/__tests__/.*|(\\.|/)(unit|e2e|intg)\\.test)\\.ts$',
   moduleNameMapper: {