+++ services/workflows-service/src/transaction/transaction.module.ts
@@ -5,26 +5,15 @@ import { TransactionRepository } from '@/transaction/transaction.repository';
 import { TransactionService } from '@/transaction/transaction.service';
 import { TransactionControllerExternal } from '@/transaction/transaction.controller.external';
 import { PrismaModule } from '@/prisma/prisma.module';
-import { ProjectScopeService } from '@/project/project-scope.service';
-import { SentryService } from '@/sentry/sentry.service';
-import { AlertService } from '@/alert/alert.service';
-import { DataAnalyticsService } from '@/data-analytics/data-analytics.service';
-import { AlertRepository } from '@/alert/alert.repository';
-import { AlertDefinitionRepository } from '@/alert-definition/alert-definition.repository';
+import { DataAnalyticsModule } from '@/data-analytics/data-analytics.module';
+import { SentryModule } from '@/sentry/sentry.module';
+import { AlertModule } from '@/alert/alert.module';
+import { ProjectModule } from '@/project/project.module';
 
 @Module({
-  imports: [ACLModule, PrismaModule],
+  imports: [ACLModule, PrismaModule, DataAnalyticsModule, SentryModule, AlertModule, ProjectModule],
   controllers: [TransactionControllerInternal, TransactionControllerExternal],
-  providers: [
-    TransactionService,
-    TransactionRepository,
-    ProjectScopeService,
-    SentryService,
-    AlertService,
-    DataAnalyticsService,
-    AlertRepository,
-    AlertDefinitionRepository,
-  ],
+  providers: [TransactionService, TransactionRepository],
   exports: [ACLModule, TransactionService],
 })
 export class TransactionModule {}