+++ packages/common/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @ballerine/common
 
+## 0.9.60
+
+### Patch Changes
+
+- Updated button with disabled state
+
+## 0.9.59
+
+### Patch Changes
+
+- core
+
+## 0.9.58
+
+### Patch Changes
+
+- Bump
+
+## 0.9.57
+
+### Patch Changes
+
+- Updated merchant screening schema
+
+## 0.9.56
+
+### Patch Changes
+
+- version bump
+
 ## 0.9.55
 
 ### Patch Changes