+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/eslint-config
 
+## 1.1.28
+
+### Patch Changes
+
+- core
+
+## 1.1.27
+
+### Patch Changes
+
+- Bump
+
 ## 1.1.26
 
 ### Patch Changes