+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @ballerine/workflow-core
 
+## 0.6.79
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.60
+
+## 0.6.78
+
+### Patch Changes
+
+- core
+- Updated dependencies
+  - @ballerine/common@0.9.59
+
+## 0.6.77
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.58
+
+## 0.6.76
+
+### Patch Changes
+
+- Added no op event to workflow runner
+
+## 0.6.75
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.57
+
+## 0.6.74
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.56
+
 ## 0.6.73
 
 ### Patch Changes
@@ -15,6 +58,7 @@
 - version bump
 
 ## 0.6.71
+
 ## 0.6.69
 
 ### Patch Changes